Add a useSubjectJSON shared hook
