+++ packages/lib-classifier/src/hooks/index.js
@@ -8,5 +8,6 @@ export { default as useProjectPreferences } from './useProjectPreferences'
 export { default as useProjectRoles } from './useProjectRoles'
 export { default as useStores } from './useStores'
 export { default as useSubjectImage } from './useSubjectImage'
+export { default as useSubjectJSON } from './useSubjectJSON'
 export { default as useTranscriptionReductions } from './useTranscriptionReductions'
 export { default as useWorkflowSnapshot } from './useWorkflowSnapshot'
